11 de mayo

+++ src/main/java/com/proyecto/Pokedex/Servicio/EntrenadorServicio.java
@@ -29,13 +29,19 @@ public class EntrenadorServicio {
 
     // BORRAR UN ENTRENADOR
     public void borrar(Long id){
-        entrenadorRepositorio.deleteById(id);
+        if(entrenadorRepositorio.existsById(id)) {
+            entrenadorRepositorio.deleteById(id);
+        }else {
+            System.out.println("No existe el entrenador con el id: " + id);
+        }
     }
 
     // MODIFICAR UN ENTRENADOR
     public Entrenador modificar(Long id, Entrenador entrenador_modificado) {
     Entrenador entrenador_viejo = entrenadorRepositorio.findById(id).orElse(null);
+    // COMPRUEBO QUE EXISTE
     if(entrenador_viejo != null){
+        // COMPRUEBO QUE LOS NUEVOS CAMPOS NO SEAN NULL, Y SI LO SON MANTENGO LOS DATOS ANTERIORES
         if(!(entrenador_modificado.getCiudad()== null)){
             entrenador_viejo.setCiudad(entrenador_modificado.getCiudad());
         }
@@ -45,9 +51,17 @@ public class EntrenadorServicio {
         if(!(entrenador_modificado.getEdad()== null)){
             entrenador_viejo.setEdad(entrenador_modificado.getEdad());
         }
-        if(!(entrenador_modificado.))
+        if(!(entrenador_modificado.getEquipo()== null)){
+            entrenador_viejo.setEquipo(entrenador_modificado.getEquipo());
+        }
+        if(!(entrenador_modificado.getPokemon_favorito()== null)){
+            entrenador_viejo.setPokemon_favorito(entrenador_modificado.getPokemon_favorito());
+        }
+        if(!(entrenador_modificado.getEquipo()== null)){
+            entrenador_viejo.setEquipo(entrenador_modificado.getEquipo());
+        }
 
     }
-    return entrenador_viejo;
+    return entrenadorRepositorio.save(entrenador_viejo);
     }
 }
